Fix bug exporting nuclear segments when interpolated segments < min length
This follows up on the issue fixed in commit 3079864fedf4cbfce9f38fdc7869d5d755cf503c. When exporting nuclear stats, the segments are normalised. When a fixed length of 1000 was used and the individual profile lengths were >1000, interpolation could bring segments to below the minimum length. Fixed by ensuring interpolated profiles are a multiple of 1000, and greater than the longest profile in any of the exported datasets.