Fix test suite issues in testing merges.
The resetting of segments and presence of the old VennCache in
VirtualDatasets was causing problems with equals testing. The merge test
now only tests for correct cell count and cell ids, but not the contents
of the cells. The VennCache has been updated to use the correct class
for VirtualDatasets.