Fix issue exporting chart pngs at correct dimensions for single column
Existing dimension checks were based on the aspect ratio of the plot;
if the plot was already fixed aspect with lots of empty space, the
resulting png could have axis ranges set too high. This fix changes the
calculation to use underlying data ranges instead, with a buffer to
prevent plotted data up to the edge of the exported image.